Added Navbar style into the component
The Navbar no longer requires an external stylesheet. Some selectors have also been combined reducing the amount of CSS.
